+++ app/routes/problem/$pid/submit.tsx
@@ -56,7 +56,7 @@ export const action: ActionFunction = async ({ request, params }) => {
   });
 
   await s3.writeFile(`/record/${rid}`, Buffer.from(code));
-  judge.push({ rid });
+  judge.push(rid);
 
   return redirect(`/record/${rid}`);
 };